replacement of text only if searching now playing
This makes it such that the text within brackets and emojis are removed if the user is searching the currently playing song but if the user is not, the brackets will stay. This could be useful if there were many songs with the same name and the user wanted to be more specific.