Update ban.js
this is what I had to change on my bot in order to successfully ban. days if you don't know is how many days of messages from userToBan will be deleted (up to 7 I believe). I wanna make it where you can choose the number of days, but I'm not sure exactly how to without possibly breaking it.