Warn User if Invalid values entered in options
The bot should check the values entered in options.json BEFORE starting and warn the user in case of error
for example
A user does not want the playLiveStreams option enabled. And mistakenly sets the playLiveStreams value to "false" instead of false.
Before setting the playLiveStreams to true, the bot should warn the user that the playLiveStreams value is invalid and the bot is setting the value to true